Added spell system and test Freeze spell
Added a simple spell sysem. Spells act as items and can be picked up. The player has 3 spell slots and can cast them with the 1/2/3 keys. When a spell is cast the area it will affect starts blinking and the player can change the direction before confirming the cast with Enter or canceling it with Escape. Need to add functionality for replacing specific spells when picking them up and adding a cooldown on spells.