Updated camera movement
Updated the camera to track which section the player is in. The sections are determined by a "buffer" input. If the player is within buffer distance from the edge of the camera they are in a different section from the center. The camera resets when the player changes camera sections.