Updated enemies to use the 3x3 attack system
Updated enemies to use the same attack system used by weapons. They have a 3x3 grid their attack can cover and the hit and damage arrays control how much damage is done to each space and the priority of each space. The BFS function needs to be updated to account for the new possible attack ranges.