Updated the enemy inRange function and added a new room buffer stat
Updated the enemy inRange function to take the targets faction as an input. Added a new room buffer stat that controls how far away rooms must spawn from eachother. Updated the dungeon generation algorithm to spawn the play and exit first before enemies and made dungeon generation fail if the player/exit can't be placed. The map constructor will try 100 times to make a map but there needs to be a preset map to load if all else fails.