Update game loop
This is a pretty big update that tries to move the game logic into a more state machine style game loop. This allows the logic to be split into Input, Player Logic, Enemy Logic, and Animation where the screen should only be updated in the Animation segment. A lot more work needs to be done on this. Currently damage animations and spells don't work and some inputs seem to be dropped. A new animation system should be developed to go with this change.